New design completed. Main functionality is working. Should test less important functions (removing animation, chain, etc). Right now Alexa works & turns on the lights on request.

+++ lib/Animo/src/anim/animationVariable.cpp
@@ -1,6 +1,6 @@
 #include "animationVariable.h"
 
-void AnimationVariable::update(unsigned long time) {
+void AnimationVariable::update(float time) {
     this->easingFunction(this, time);
     this->updateCallback(*this);
 }
